Break interfaceHelpers/Pages.jsx into multiple files
