+++ src/components/Interfaces/Documentation.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import PagesInterface from './interfaceHelpers/Pages';
+import MarkdownFiles from '../Markdown/Files';
 import docs from '../../resources/docs';
 
 const DocumentationInterface = props => (
-  <PagesInterface pages={docs} homePage="welcome" whiteList="docs" {...props} />
+  <MarkdownFiles pages={docs} homePage="welcome" whiteList="docs" {...props} />
 );
 
 export default DocumentationInterface;
